Remove the profileDetails service and hook from home page and added it under profileDetails folder
