+++ src/app/screens/PeerlyScreens/HomeScreen/home.hooks.ts
@@ -4,28 +4,12 @@ import {AxiosError} from 'axios';
 import toast from '../../../utils/toast';
 
 import {
-  getProfileDetails,
   getTopUsersList,
   getActiveUsersList,
   getAppreciationList,
 } from '../../../services/PeerlyServices/home';
 import {APIError} from '../types';
 import {GetAppreciationListRequest} from '../../../services/PeerlyServices/home/types';
-
-export function useGetProfileDetails() {
-  const {data, isLoading, isFetching, isSuccess, isError} = useQuery({
-    queryKey: ['profile_icon'],
-    queryFn: getProfileDetails,
-    onError: (error: AxiosError<APIError>) => {
-      if (error.response?.data.message) {
-        toast(error.response.data.message, 'error');
-      } else {
-        toast('Something went wrong while fetching profile details', 'error');
-      }
-    },
-  });
-  return {data: data?.data, isLoading, isFetching, isSuccess, isError};
-}
 
 export function useGetTopUsersList() {
   const {data, isLoading, isFetching, isSuccess, isError} = useQuery({
