+++ src/app/screens/PeerlyScreens/ProfileDetailScreen/index.tsx
@@ -1,6 +1,7 @@
 import {Text, View} from 'react-native';
+import {useGetProfileDetails} from './profile.hooks';
 
-export default function ProfileScreen() {
+export default function ProfileScreen({navigate}) {
   return (
     <View>
       <Text>ProfileScreen</Text>
