--- src/app/screens/PeerlyScreens/ProfileScreen/profile.hooks.ts
@@ -1,1 +1,0 @@
-export function demohook() {}
